+++ freud/box.pyx
@@ -16,6 +16,7 @@ import freud.util
 
 cimport numpy as np
 from cpython.object cimport Py_EQ, Py_NE
+from cython.operator cimport dereference
 from libcpp cimport bool as cpp_bool
 
 cimport freud._box
@@ -541,15 +542,18 @@ cdef class Box:
         return np.asarray(distances)
 
     def contains(self, points):
-        r"""Returns boolean array (mask) corresponding to point membership in a box.
+        r"""Compute a boolean array (mask) corresponding to point membership in a box.
 
-        This calculation computes particle membership based on conventions defined by :class:`Box`, ignoring periodicity.
-        This means that in a cubic (3D) box with dimensions ``L``, particles would be considered inside the box if their coordinates are between
-        ``[-L/2, L/2]``.
-        Particles laying at a coordinate such as ``[0, L, 0]`` would be considered outside the box.
-        More information about coordinate conventions can be found `here
-        <https://freud.readthedocs.io/en/latest/gettingstarted/examples/module_intros/box.Box.html?highlight=origin#Using-boxes>`__
-        and `here <https://freud.readthedocs.io/en/latest/gettingstarted/tutorial/periodic.html?highlight=origin#periodic-boundary-conditions>`__.
+        This calculation computes particle membership based on conventions
+        defined by :class:`Box`, ignoring periodicity. This means that in a
+        cubic (3D) box with dimensions ``L``, particles would be considered
+        inside the box if their coordinates are between ``[-L/2, L/2]``.
+        Particles laying at a coordinate such as ``[0, L, 0]`` would be
+        considered outside the box.  More information about coordinate
+        conventions can be found in the documentation on `Using boxes
+        <https://freud.readthedocs.io/en/latest/gettingstarted/examples/module_intros/box.Box.html#Using-boxes>`__
+        and `periodic boundary conditions
+        <https://freud.readthedocs.io/en/latest/gettingstarted/tutorial/periodic.html#periodic-boundary-conditions>`__.
 
         Example::
 
@@ -565,8 +569,8 @@ cdef class Box:
 
         Returns:
             :math:`\left(N, \right)` :class:`numpy.ndarray`:
-                Array of booleans, where `True` corresponds to points within the box,
-                and `False` corresponds to points outside the box.
+                Array of booleans, where ``True`` corresponds to points within
+                the box, and ``False`` corresponds to points outside the box.
         """  # noqa: E501
 
         points = freud.util._convert_array(
@@ -574,14 +578,14 @@ cdef class Box:
 
         cdef:
             const float[:, ::1] l_points = points
-            size_t n_all_points = points.shape[0]
+            size_t n_points = points.shape[0]
 
         contains_mask = freud.util._convert_array(
-            np.ones(n_all_points), dtype=bool)
+            np.ones(n_points), dtype=bool)
         cdef cpp_bool[::1] l_contains_mask = contains_mask
 
         self.thisptr.contains(
-            <vec3[float]*> &l_points[0, 0], n_all_points,
+            <vec3[float]*> &l_points[0, 0], n_points,
             <cpp_bool*> &l_contains_mask[0])
 
         return np.array(l_contains_mask).astype(bool)
@@ -684,17 +688,19 @@ cdef class Box:
     def __str__(self):
         return repr(self)
 
-    def _eq(self, other):
-        return self.to_dict() == other.to_dict()
-
     def __richcmp__(self, other, int op):
         r"""Implement all comparisons for Cython extension classes"""
-        if op == Py_EQ:
-            return self._eq(other)
-        if op == Py_NE:
-            return not self._eq(other)
-        else:
-            raise NotImplementedError("This comparison is not implemented")
+        cdef Box c_other
+        try:
+            c_other = <Box?> other
+            if op == Py_EQ:
+                return dereference(self.thisptr) == dereference(c_other.thisptr)
+            if op == Py_NE:
+                return dereference(self.thisptr) != dereference(c_other.thisptr)
+        except TypeError:
+            # Cython cast to Box failed
+            pass
+        return NotImplemented
 
     def __mul__(arg1, arg2):
         # Note Cython treats __mul__ and __rmul__ as one operation, so